feat: separate converters
